Fix some types. Fix compiler warnings on 64bit systems regarding the threadid in log/monitor. It is now stored in the client structure and it is made sure that no collisions can occur on 64bit systems.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5838 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-cccshare.c
@@ -636,7 +636,7 @@ int32_t equal_providers(struct cc_card *card1, struct cc_card *card2) {
 /**
  * Adds a new card to a cardlist.
  */
-int32_t add_card_to_serverlist(LLIST *cardlist, struct cc_card *card, int free_card) {
+int32_t add_card_to_serverlist(LLIST *cardlist, struct cc_card *card, int8_t free_card) {
 
     int32_t modified = 0;
     LL_ITER it = ll_iter_create(cardlist);
